requirements.txt
pytorch=1.2.0
numpy=1.17.3
pydicom=1.2.2
scikit-learn=0.21.3
torchvision=0.4.0
pyyaml=5.1.2
pandas=0.25.3
